fix(array): indexOf/lastIndexOf i64 return, frozen/non-writable push/unshift guards (fixes #5589) (#5794)
Three categories of test262 Array failures addressed:

1. indexOf/lastIndexOf i32 overflow: return type widened to i64 in search.rs
   and all three codegen call sites (strings.rs, lower_array_method.rs,
   misc_methods.rs) so indices >= 2^31 survive without sign-truncation.

2. Frozen/non-writable-length push guards: guards added to every push dispatch
   path — real_array_mutator "push" arm (generic.rs), common_methods.rs
   dispatch_common "push" arm, handle_methods.rs GC_TYPE_ARRAY "push" arm,
   and the static NativeMethodCall push_single / push codegen path via the new
   js_array_push_guard runtime function (push_pop.rs + arrays.rs).
   The static path previously skipped mutation checks for 0-arg push entirely.

3. Frozen/non-writable-length unshift(0-arg) guards: the static codegen path
   for 0-arg unshift (lower_array_method.rs) now routes through
   js_array_unshift_variadic(arr, null, 0) instead of directly calling
   js_array_length, so the runtime guards in push_pop.rs fire.
   The real_array_mutator "unshift" arm (generic.rs) also routes zero-arg
   calls through js_array_unshift_variadic for the same reason.

Verified 12/12 frozen-array mutation tests and 10/10 indexOf/lastIndexOf
overflow tests all pass on the rebuilt binary.

+++ crates/perry-runtime/src/array/search.rs
@@ -114,7 +114,7 @@ pub extern "C" fn js_array_indexOf_jsvalue(
     value: f64,
     from_index: f64,
     has_from: i32,
-) -> i32 {
+) -> i64 {
     let arr = normalize_array_receiver(arr);
     if arr.is_null() {
         return -1;
@@ -137,7 +137,7 @@ pub extern "C" fn js_array_indexOf_jsvalue(
             if crate::value::js_jsvalue_equals(crate::typedarray::js_typed_array_get(ta, i), value)
                 == 1
             {
-                return i;
+                return i as i64;
             }
         }
         return -1;
@@ -169,7 +169,7 @@ pub extern "C" fn js_array_indexOf_jsvalue(
                 }
             };
             if crate::value::js_jsvalue_equals(element, value) == 1 {
-                return i as i32;
+                return i;
             }
         }
         -1
@@ -187,7 +187,7 @@ pub extern "C" fn js_array_last_index_of_jsvalue(
     value: f64,
     from_index: f64,
     has_from: i32,
-) -> i32 {
+) -> i64 {
     let arr = normalize_array_receiver(arr);
     if arr.is_null() {
         return -1;
@@ -221,7 +221,7 @@ pub extern "C" fn js_array_last_index_of_jsvalue(
                 value,
             ) == 1
             {
-                return i as i32;
+                return i as i64;
             }
             i -= 1;
         }
@@ -271,7 +271,7 @@ pub extern "C" fn js_array_last_index_of_jsvalue(
                 }
             };
             if crate::value::js_jsvalue_equals(element, value) == 1 {
-                return i as i32;
+                return i;
             }
             i -= 1;
         }
